Notification functional. Updated logic for fetching new posts. Notification database. Additional authrorization measures to make editing and deleting posts more secure. Backend console logs for information.

+++ band-server/src/routes/postRoutes.js
@@ -5,45 +5,59 @@ import dayjs from 'dayjs';
 
 const router = express.Router();
 
-// Get all posts
-router.get('/:sort', (req, res) => {
-  // const getPosts = db.prepare(`SELECT * FROM posts WHERE user_id = ?`);
-  // const posts = getPosts.all(req.userId);
-  const {sort} = req.params; // 0: Date (Desc), 1: Date (Asc), 2: Title (Desc), 3: Title (Asc)
-  let type = 'title';
-  let order = 'DESC';
 
-  if (sort === '0') {
-    type = 'date';
-    order = 'DESC'
-  } else if (sort === '1') {
-    type = 'date';
-    order = 'ASC'
-  } else if (sort === '2') {
-    type = 'title';
-    order = 'DESC'
-  } else if (sort === '3') {
-    type = 'title';
-    order = 'ASC'
+router.get('/:type', (req, res) => {
+
+  // const {sort, mostRecentPID} = req.params; // 0: Date (Desc), 1: Date (Asc), 2: Title (Desc), 3: Title (Asc)
+  // const {pid, sort} = req.query;
+
+  // either 'all' or 'comments'
+  const { type } = req.params; 
+  // pid: Associated pid. If -1, then fetch all posts. Otherwise, fetch posts newer than pid
+  // sort: 0: Date (Desc), 1: Date (Asc), 2: Title (Desc), 3: Title (Asc)
+  const { pid, sort } = req.query; 
+
+  if (type === 'all') {
+    // console.log('get all')
+    let sortOn = 'title';
+    let order = 'DESC';
+
+    if (sort === '0') {
+      sortOn = 'date';
+      order = 'DESC'
+    } else if (sort === '1') {
+      sortOn = 'date';
+      order = 'ASC'
+    } else if (sort === '2') {
+      sortOn = 'title';
+      order = 'DESC'
+    } else if (sort === '3') {
+      sortOn = 'title';
+      order = 'ASC'
+    }
+
+    const getPosts = db.prepare(`SELECT * FROM posts WHERE pid > ? ORDER BY ${sortOn} ${order}`);
+    const posts = getPosts.all(pid);
+    // console.log('\n\n\n=============== Posts:')
+    // console.log(posts);
+    console.log(`Get all posts for user ${req.userId} with search ${sortOn}/${order} (${sort})`);
+    res.json(posts);
+  } else if (type === 'comments') {
+    // console.log('get comments');
+    const getComments = db.prepare(`SELECT * FROM posts WHERE parent = ? ORDER BY date DESC`);
+    const comments = getComments.all(pid);
+
+    // console.log(`comments with pid parent ${pid}`);
+    // console.log(comments);
+
+    console.log(`Get comments for user ${req.userId} on post ${pid}`);
+
+    res.json(comments);
   }
 
-  // console.log(`ORDER BY ${type} ${order}`);
 
-  const getPosts = db.prepare(`SELECT * FROM posts ORDER BY ${type} ${order}`);
-  const posts = getPosts.all();
-
-  // console.log(posts);
-  res.json(posts);
 })
 
-// Get all posts with parent != null (i.e., comment), and parent = req.body.pid;
-router.get('/comments/:pid', (req, res) => {
-  const {pid} = req.params;
-  const getComments = db.prepare(`SELECT * FROM posts WHERE parent = ? ORDER BY date DESC`);
-  const comments = getComments.all(pid);
-
-  res.json(comments);
-})
 
 // Create a new post
 router.post('/', (req, res) => {
@@ -60,9 +74,9 @@ router.post('/', (req, res) => {
   // req.userId is inserted into the request from authMiddleware
   const result = insertPost.run(req.userId, username, title, content, tags, date, 1, parent);
 
-  res.json({pid: result.lastInsertRowid, message:"success"}) // send more info back?
+  console.log(`Creating a post from user ${req.userId} ${parent && `(comment on post ${parent})`}`);
 
-
+  res.json({pid: result.lastInsertRowid, message:"Post success"}) // send more info back?
 })
 
 // Update/edit a new post
@@ -71,8 +85,20 @@ router.put('/:id', (req, res) => {
   const {id} = req.params; // i.e., posts/2
   // const {page} = req.query // i.e., posts/?page=1
 
+  const postOwner = db.prepare(`SELECT * FROM posts WHERE pid = ?`);
+  const postOwnerData = postOwner.all(id);
+  console.log(postOwnerData);
+
+  if (postOwnerData[0].user_id !== req.userId) {
+    console.log(`User ${req.userId} tried to edit post owned by ${postOwnerData[0].user_id}`);
+    res.sendStatus(401);
+    return;
+  }
+
   const updatedPost = db.prepare(`UPDATE posts SET title = ?, content = ?, tags = ? WHERE pid = ?`);
   updatedPost.run(title, content, tagString, id);
+
+  console.log(`Updating post from user ${req.userId} on post ${id}`);
 
   res.json({message:"post edited"})
 })
@@ -82,8 +108,20 @@ router.delete('/:id', (req, res) => {
   const {id} = req.params;
   const userId = req.userId;
 
+  const postOwner = db.prepare(`SELECT * FROM posts WHERE pid = ?`);
+  const postOwnerData = postOwner.all(id);
+  // console.log(postOwnerData);
+
+  if (postOwnerData[0].user_id !== req.userId) {
+    console.log(`User ${req.userId} tried to delete post owned by ${postOwnerData[0].user_id}`);
+    res.sendStatus(401);
+    return;
+  }
+
   const deletePosts = db.prepare(`DELETE FROM posts WHERE pid = ? AND user_id = ?`);
   deletePosts.run(id, userId);
+
+  console.log(`Deleting post from user ${req.userId} on post ${id}`);
 
   res.json({message:"deleted"})
